Refactor test helpers and add new selectAtIndex frost-select helper
